# src/backend/REST-API/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Eine API zur Verbindung mit unserer Supabase-Datenbank
  title: LaendleGuessr
  version: 1.0.0
  description: Diese OpenAPI-Spezifikation beschreibt die REST-API für das LaendleGuessr-Projekt.
servers:
  - url: /
    description: Lokaler Server
paths:
...
  /user:
    get:
      operationId: all_users_get
      responses:
        "200":
          description: Erfolgreiche Antwort
        "400":
          description: Ungültige Eingabe
        "503":
          description: Service nicht verfügbar
      summary: Alle User anzeigen
      tags:
        - user
      x-openapi-router-controller: openapi_server.controllers.user_controller
    post:
      operationId: user_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        "200":
          description: Erfolgreiche Antwort
        "400":
          description: Ungültige Eingabe
        "503":
          description: Service nicht verfügbar
      summary: Neuen User hinzufügen
      tags:
        - user
      x-openapi-router-controller: openapi_server.controllers.user_controller
  /user/{id}:
    get:
      operationId: user_get
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Erfolgreiche Antwort
        "400":
          description: Ungültige Eingabe
        "503":
          description: Service nicht verfügbar
      summary: User mit bestimmter ID anzeigen
      tags:
        - user
      x-openapi-router-controller: openapi_server.controllers.user_controller
    put:
      operationId: user_update
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        "200":
          description: Erfolgreiche Antwort
        "400":
          description: Ungültige Eingabe
        "503":
          description: Service nicht verfügbar
      summary: User bearbeiten
      tags:
        - user
      x-openapi-router-controller: openapi_server.controllers.user_controller
  /user/login:
    post:
      summary: Benutzeranmeldung
      operationId: user_login
      x-openapi-router-controller: openapi_server.controllers.user_controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: maxmustermann
                password:
                  type: string
                  example: geheim123
              required:
                - username
                - password
      responses:
        '200':
          description: Erfolgreich angemeldet
          content:
            application/json:
              schema:
                type: object
                properties:
                  nachricht:
                    type: string
                    example: Anmeldung erfolgreich
                  uid:
                    type: string
                  username:
                    type: string
                  coins:
                    type: integer
                  admin:
                    type: boolean
                  city:
                    type: integer
        '400':
          description: Ungültige Anfrage
          content:
            application/json:
              schema:
                type: object
                properties:
                  nachricht:
                    type: string
                    example: Benutzername und Passwort erforderlich
        '401':
          description: Ungültige Anmeldedaten
          content:
            application/json:
              schema:
                type: object
                properties:
                  nachricht:
                    type: string
                    example: Ungültiger Benutzername oder Passwort
  /all_time_stats/{uid}:
    get:
      operationId: get_all_time_stats
      parameters:
        - in: path
          name: uid
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Erfolgreiche Antwort
        "400":
          description: Ungültige Eingabe
        "503":
          description: Service nicht verfügbar
      summary: Gibt die Allzeit-Statistiken eines Users zurück
      tags:
        - user
      x-openapi-router-controller: openapi_server.controllers.user_controller
...
components:
  schemas:
...
    user:
      title: user
      description: Repräsentiert einen Benutzer im System.
      required:
        - username
        - password
        - coins
        - admin
        - city
      properties:
        username:
          type: string
          title: Benutzername
          minLength: 3
          maxLength: 30
          description: Der eindeutige Benutzername
        password:
          type: string
          title: Passwort
          description: Das Passwort des Benutzers
        coins:
          type: integer
          title: Coins
          minimum: 0
          description: Die Anzahl der Münzen, die der Benutzer besitzt
        admin:
          type: boolean
          title: Admin
          description: Gibt an, ob der Benutzer Administratorrechte hat
        city:
          type: integer
          title: Stadt
          enum:
            - 0  
            - 1  
            - 2  
            - 3  
            - 4  
          description: Die Stadt-ID, der der Benutzer zugeordnet ist
      example:
        username: keremderkomet
        password: MeinSicheresPasswort123
        coins: 150
        admin: true
        city: 1
...